backend/requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
# pywin32>=300  # Windows only
pypdf2>=3.0.0
pycryptodome>=3.15.0
pillow>=9.0.0
psutil>=5.9.0
python-barcode>=0.15.1
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
flask-login>=0.6.2
flask-jwt-extended>=4.5.2
python-socketio>=5.9.0
bcrypt>=4.0.1
pyjwt>=2.8.0